# mobile/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "GH Trust MFB API",
  "description": "GH Trust International Ltd \u2014 Microfinance Banking API",
  "version": "0.2.0"
 },
 "paths": {
...
  "/api/v1/admin/loans/products": {
   "get": {
    "tags": [
     "Admin \u2014 Loans"
    ],
    "summary": "Admin List Products",
    "operationId": "admin_list_products_api_v1_admin_loans_products_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/LoanProductResponse"
         },
         "type": "array",
         "title": "Response Admin List Products Api V1 Admin Loans Products Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   },
   "post": {
    "tags": [
     "Admin \u2014 Loans"
    ],
    "summary": "Create loan product",
    "description": "Created switched off. Publish an approval workflow for it, then switch it on.",
    "operationId": "admin_create_product_api_v1_admin_loans_products_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CreateLoanProductRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/LoanProductResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
...
  "/api/v1/admin/loans/loans": {
   "get": {
    "tags": [
     "Admin \u2014 Loans"
    ],
    "summary": "Loan book",
    "operationId": "admin_list_loans_api_v1_admin_loans_loans_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "$ref": "#/components/schemas/LoanStatus"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "name": "customer_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Customer Id"
      }
     },
     {
      "name": "search",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 100
        },
        {
         "type": "null"
        }
       ],
       "description": "Borrower name, phone, account or exact BVN",
       "title": "Search"
      },
      "description": "Borrower name, phone, account or exact BVN"
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_LoanResponse_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/loans/me/applications": {
   "get": {
    "tags": [
     "Loans"
    ],
    "summary": "List My Applications",
    "operationId": "list_my_applications_api_v1_loans_me_applications_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "$ref": "#/components/schemas/ApplicationStatus"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 100,
       "minimum": 1,
       "default": 20,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_LoanApplicationSummaryResponse_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "Loans"
    ],
    "summary": "Create My Application",
    "operationId": "create_my_application_api_v1_loans_me_applications_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CreateApplicationRequest"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/LoanApplicationDetailResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/loans/me/applications/{application_id}": {
   "get": {
    "tags": [
     "Loans"
    ],
    "summary": "Get My Application",
    "operationId": "get_my_application_api_v1_loans_me_applications__application_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "application_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Application Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/LoanApplicationDetailResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "patch": {
    "tags": [
     "Loans"
    ],
    "summary": "Update My Application Step",
    "operationId": "update_my_application_step_api_v1_loans_me_applications__application_id__patch",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "application_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Application Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UpdateApplicationStepRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/LoanApplicationDetailResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/loans/me/applications/{application_id}/documents/{document_type}": {
   "post": {
    "tags": [
     "Loans"
    ],
    "summary": "Upload My Application Document",
    "operationId": "upload_my_application_document_api_v1_loans_me_applications__application_id__documents__document_type__post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "application_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Application Id"
      }
     },
     {
      "name": "document_type",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Document Type"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_upload_my_application_document_api_v1_loans_me_applications__application_id__documents__document_type__post"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/LoanApplicationDetailResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/loans/me/applications/{application_id}/submit": {
   "post": {
    "tags": [
     "Loans"
    ],
    "summary": "Submit My Application",
    "operationId": "submit_my_application_api_v1_loans_me_applications__application_id__submit_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "application_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Application Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/LoanApplicationDetailResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/loans/me/loans": {
   "get": {
    "tags": [
     "Loans"
    ],
    "summary": "List My Loans",
    "operationId": "list_my_loans_api_v1_loans_me_loans_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "$ref": "#/components/schemas/LoanStatus"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 100,
       "minimum": 1,
       "default": 20,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_LoanResponse_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/loans/me/loans/{loan_id}": {
   "get": {
    "tags": [
     "Loans"
    ],
    "summary": "Get My Loan",
    "description": "Loan with its full repayment schedule and payment history.",
    "operationId": "get_my_loan_api_v1_loans_me_loans__loan_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "loan_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Loan Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/LoanDetailResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/loans/me/loans/{loan_id}/repayments": {
   "post": {
    "tags": [
     "Loans"
    ],
    "summary": "Repay from wallet",
    "description": "Debit the customer's wallet and apply it to the loan (oldest installment first, interest before principal). Requires an `Idempotency-Key` header: retrying with the same key never charges twice.",
    "operationId": "repay_my_loan_api_v1_loans_me_loans__loan_id__repayments_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "loan_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Loan Id"
      }
     },
     {
      "name": "Idempotency-Key",
      "in": "header",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Idempotency-Key"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CustomerRepayRequest"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/LoanRepaymentResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "ApplicationChannel": {
    "type": "string",
    "enum": [
     "web",
     "branch",
     "mobile"
    ],
    "title": "ApplicationChannel"
   },
...
   "ApplicationStatus": {
    "type": "string",
    "enum": [
     "draft",
     "submitted",
     "under_review",
     "documents_incomplete",
     "approved",
     "offer_sent",
     "offer_accepted",
     "product_gate_pending",
     "processing_fee_paid",
     "ready_to_disburse",
     "disbursed",
     "rejected",
     "withdrawn",
     "expired"
    ],
    "title": "ApplicationStatus"
   },
...
   "Body_upload_my_application_document_api_v1_loans_me_applications__application_id__documents__document_type__post": {
    "properties": {
     "file": {
      "type": "string",
      "format": "binary",
      "title": "File"
     }
    },
    "type": "object",
    "required": [
     "file"
    ],
    "title": "Body_upload_my_application_document_api_v1_loans_me_applications__application_id__documents__document_type__post"
   },
...
   "CreateApplicationRequest": {
    "properties": {
     "product_code": {
      "type": "string",
      "maxLength": 50,
      "minLength": 1,
      "title": "Product Code"
     },
     "channel": {
      "$ref": "#/components/schemas/ApplicationChannel",
      "default": "web"
     }
    },
    "type": "object",
    "required": [
     "product_code"
    ],
    "title": "CreateApplicationRequest"
   },
   "CreateLoanProductRequest": {
    "properties": {
     "code": {
      "type": "string",
      "pattern": "^[a-z][a-z0-9_]{2,49}$",
      "title": "Code",
      "description": "Permanent ID, e.g. `agric_loan`"
     },
     "name": {
      "type": "string",
      "maxLength": 150,
      "minLength": 3,
      "title": "Name"
     },
     "description": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 500
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "interest_rate_pct_monthly": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 20.0,
        "exclusiveMinimum": 0.0
       },
       {
        "type": "string"
       }
      ],
      "title": "Interest Rate Pct Monthly"
     },
     "processing_fee_pct": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 20.0,
        "minimum": 0.0
       },
       {
        "type": "string"
       }
      ],
      "title": "Processing Fee Pct",
      "default": "0"
     },
     "interest_method": {
      "$ref": "#/components/schemas/InterestMethod",
      "default": "flat"
     },
     "max_tenure_days": {
      "anyOf": [
       {
        "type": "integer",
        "maximum": 1825.0,
        "minimum": 30.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Tenure Days"
     },
     "default_penalty_pct_daily": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 5.0,
        "minimum": 0.0
       },
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Default Penalty Pct Daily"
     },
     "repayment_cadence_options": {
      "items": {
       "$ref": "#/components/schemas/RepaymentCadence"
      },
      "type": "array",
      "minItems": 1,
      "title": "Repayment Cadence Options"
     },
     "required_document_types": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "minItems": 1,
      "title": "Required Document Types"
     },
     "workflow_steps": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "minItems": 3,
      "title": "Workflow Steps"
     }
    },
    "type": "object",
    "required": [
     "code",
     "name",
     "interest_rate_pct_monthly",
     "repayment_cadence_options",
     "required_document_types",
     "workflow_steps"
    ],
    "title": "CreateLoanProductRequest",
    "description": "A new product, created switched off. Steps, documents and cadences are limited to the\nones the customer app and servicing engine understand; switching it on requires a\npublished approval workflow."
   },
...
   "CustomerRepayRequest": {
    "properties": {
     "amount": {
      "anyOf": [
       {
        "type": "number",
        "exclusiveMinimum": 0.0
       },
       {
        "type": "string"
       }
      ],
      "title": "Amount"
     },
     "transaction_pin": {
      "type": "string",
      "maxLength": 4,
      "minLength": 4,
      "pattern": "^\\d+$",
      "title": "Transaction Pin",
      "description": "Customer's 4-digit transaction PIN"
     }
    },
    "type": "object",
    "required": [
     "amount",
     "transaction_pin"
    ],
    "title": "CustomerRepayRequest"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "InterestMethod": {
    "type": "string",
    "enum": [
     "flat",
     "reducing_balance"
    ],
    "title": "InterestMethod"
   },
...
   "LoanApplicationDetailResponse": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "customer_id": {
      "type": "string",
      "title": "Customer Id"
     },
     "product_code": {
      "type": "string",
      "title": "Product Code"
     },
     "product_name": {
      "type": "string",
      "title": "Product Name"
     },
     "status": {
      "$ref": "#/components/schemas/ApplicationStatus"
     },
     "channel": {
      "$ref": "#/components/schemas/ApplicationChannel"
     },
     "step": {
      "type": "integer",
      "title": "Step"
     },
     "total_steps": {
      "type": "integer",
      "title": "Total Steps"
     },
     "branch": {
      "type": "string",
      "title": "Branch"
     },
     "requested_amount": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Requested Amount"
     },
     "approved_amount": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Approved Amount"
     },
     "repayment_cadence": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Repayment Cadence"
     },
     "approved_tenure_months": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Approved Tenure Months"
     },
     "tenure_months": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Tenure Months",
      "description": "Effective tenure: staff-approved, else what the customer entered"
     },
     "universal_form": {
      "type": "object",
      "title": "Universal Form"
     },
     "product_data": {
      "type": "object",
      "title": "Product Data"
     },
     "submitted_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Submitted At"
     },
     "assigned_officer_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Assigned Officer Id"
     },
     "rejection_reason": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Rejection Reason"
     },
     "guarantors": {
      "items": {
       "$ref": "#/components/schemas/GuarantorResponse"
      },
      "type": "array",
      "title": "Guarantors"
     },
     "collaterals": {
      "items": {
       "$ref": "#/components/schemas/CollateralResponse"
      },
      "type": "array",
      "title": "Collaterals"
     },
     "documents": {
      "items": {
       "$ref": "#/components/schemas/ApplicationDocumentResponse"
      },
      "type": "array",
      "title": "Documents"
     },
     "document_checklist": {
      "items": {
       "$ref": "#/components/schemas/DocumentChecklistItem"
      },
      "type": "array",
      "title": "Document Checklist"
     }
    },
    "type": "object",
    "required": [
     "id",
     "customer_id",
     "product_code",
     "product_name",
     "status",
     "channel",
     "step",
     "total_steps",
     "branch",
     "requested_amount",
     "approved_amount",
     "repayment_cadence",
     "universal_form",
     "product_data",
     "submitted_at",
     "assigned_officer_id",
     "rejection_reason",
     "guarantors",
     "collaterals",
     "documents",
     "document_checklist"
    ],
    "title": "LoanApplicationDetailResponse"
   },
...
   "LoanProductResponse": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "code": {
      "type": "string",
      "title": "Code"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "is_active": {
      "type": "boolean",
      "title": "Is Active"
     },
     "processing_fee_pct": {
      "type": "string",
      "title": "Processing Fee Pct"
     },
     "interest_rate_pct_monthly": {
      "type": "string",
      "title": "Interest Rate Pct Monthly"
     },
     "max_tenure_days": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Tenure Days"
     },
     "default_penalty_pct_daily": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Default Penalty Pct Daily"
     },
     "repayment_cadence_options": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Repayment Cadence Options"
     },
     "required_document_types": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Required Document Types"
     },
     "workflow_steps": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Workflow Steps"
     },
     "eligibility_rules": {
      "type": "object",
      "title": "Eligibility Rules"
     }
    },
    "type": "object",
    "required": [
     "id",
     "code",
     "name",
     "description",
     "is_active",
     "processing_fee_pct",
     "interest_rate_pct_monthly",
     "max_tenure_days",
     "default_penalty_pct_daily",
     "repayment_cadence_options",
     "required_document_types",
     "workflow_steps",
     "eligibility_rules"
    ],
    "title": "LoanProductResponse"
   },
...
   "LoanRepaymentResponse": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "loan_id": {
      "type": "string",
      "title": "Loan Id"
     },
     "amount": {
      "type": "string",
      "title": "Amount"
     },
     "principal_amount": {
      "type": "string",
      "title": "Principal Amount"
     },
     "interest_amount": {
      "type": "string",
      "title": "Interest Amount"
     },
     "channel": {
      "$ref": "#/components/schemas/RepaymentChannel"
     },
     "reference": {
      "type": "string",
      "title": "Reference"
     },
     "paid_at": {
      "type": "string",
      "format": "date-time",
      "title": "Paid At"
     },
     "note": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Note"
     }
    },
    "type": "object",
    "required": [
     "id",
     "loan_id",
     "amount",
     "principal_amount",
     "interest_amount",
     "channel",
     "reference",
     "paid_at"
    ],
    "title": "LoanRepaymentResponse"
   },
...
   "LoanStatus": {
    "type": "string",
    "enum": [
     "active",
     "overdue",
     "completed",
     "written_off"
    ],
    "title": "LoanStatus"
   },
...
   "Page_LoanApplicationSummaryResponse_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/LoanApplicationSummaryResponse"
      },
      "type": "array",
      "title": "Items"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "limit": {
      "type": "integer",
      "title": "Limit"
     },
     "offset": {
      "type": "integer",
      "title": "Offset"
     }
    },
    "type": "object",
    "required": [
     "items",
     "total",
     "limit",
     "offset"
    ],
    "title": "Page[LoanApplicationSummaryResponse]"
   },
   "Page_LoanResponse_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/LoanResponse"
      },
      "type": "array",
      "title": "Items"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "limit": {
      "type": "integer",
      "title": "Limit"
     },
     "offset": {
      "type": "integer",
      "title": "Offset"
     }
    },
    "type": "object",
    "required": [
     "items",
     "total",
     "limit",
     "offset"
    ],
    "title": "Page[LoanResponse]"
   },
...
   "RepaymentCadence": {
    "type": "string",
    "enum": [
     "daily",
     "weekly",
     "monthly",
     "salary_date"
    ],
    "title": "RepaymentCadence"
   },
   "RepaymentChannel": {
    "type": "string",
    "enum": [
     "wallet",
     "bank_transfer",
     "cash",
     "remita",
     "other"
    ],
    "title": "RepaymentChannel"
   },
...
   "UpdateApplicationStepRequest": {
    "properties": {
     "step": {
      "type": "integer",
      "minimum": 1.0,
      "title": "Step"
     },
     "total_steps": {
      "type": "integer",
      "minimum": 1.0,
      "title": "Total Steps"
     },
     "universal_form": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/UniversalFormData"
       },
       {
        "type": "null"
       }
      ]
     },
     "product_data": {
      "anyOf": [
       {
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Product Data"
     },
     "guarantors": {
      "anyOf": [
       {
        "items": {
         "$ref": "#/components/schemas/GuarantorInput"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Guarantors"
     },
     "collaterals": {
      "anyOf": [
       {
        "items": {
         "$ref": "#/components/schemas/CollateralInput"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Collaterals"
     }
    },
    "type": "object",
    "required": [
     "step",
     "total_steps"
    ],
    "title": "UpdateApplicationStepRequest"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 }
}